+++ tests/unit/test_convert.py
@@ -1,11 +1,12 @@
-from scanapi.convert import convert
+from scanapi.convert import openapi_to_scanapi
+from scanapi.settings import settings
 import logging
 from pytest import mark, raises
 
 
-@mark.describe("convert")
-@mark.describe("convert")
-class TestConvert:
+@mark.describe("openapi_to_scanapi")
+@mark.describe("openapi_to_scanapi")
+class TestOpenAPIToScanAPI:
     @mark.context("when called with an openapi_path")
     @mark.it("should call the resolving parser on the openapi_path")
     def test_call_resolving_parser_on_path(self, mocker):
@@ -21,7 +22,11 @@ class TestConvert:
         mocker.patch("scanapi.convert.yaml.dump")
         mocker.patch("builtins.open", mocker.mock_open())
 
-        convert("openapi.json", "$BASE_URL", "scanapi.yaml")
+        settings["input_path"] = "openapi.json"
+        settings["base_url"] = "$BASE_URL"
+        settings["output_path"] = "scanapi.yaml"
+
+        openapi_to_scanapi()
         mock_resolving_parser.assert_called_once_with("openapi.json")
 
     @mark.context("when the convertion runs successfully")
@@ -39,13 +44,18 @@ class TestConvert:
         converter_mock.convert.return_value = {"test": "test"}, set()
 
         mocker.patch(
-            "scanapi.convert.OpenAPIConverter", return_value=converter_mock
+            "scanapi.convert.OpenAPIToScanAPIConverter",
+            return_value=converter_mock,
         )
 
         yaml_dump_mock = mocker.patch("scanapi.convert.yaml.dump")
         open_mock = mocker.patch("builtins.open", mocker.mock_open())
 
-        convert("openapi.json", "$BASE_URL", "test.yaml")
+        settings["input_path"] = "openapi.json"
+        settings["base_url"] = "$BASE_URL"
+        settings["output_path"] = "test.yaml"
+
+        openapi_to_scanapi()
         open_mock.assert_called_once_with("test.yaml", "w")
         yaml_dump_mock.assert_called_once_with(
             {"test": "test"},
@@ -73,13 +83,18 @@ class TestConvert:
         )
 
         mocker.patch(
-            "scanapi.convert.OpenAPIConverter", return_value=converter_mock
+            "scanapi.convert.OpenAPIToScanAPIConverter",
+            return_value=converter_mock,
         )
 
         mocker.patch("scanapi.convert.yaml.dump")
         mocker.patch("builtins.open", mocker.mock_open())
 
-        convert("openapi.json", "$BASE_URL", "test.yaml")
+        settings["input_path"] = "openapi.json"
+        settings["base_url"] = "$BASE_URL"
+        settings["output_path"] = "test.yaml"
+
+        openapi_to_scanapi()
         captured = capsys.readouterr()
 
         assert "The following variables were created" in captured.out
@@ -92,11 +107,10 @@ class TestConvert:
     def test_handles_invalid_yaml_file(self, caplog):
         with caplog.at_level(logging.ERROR):
             with raises(SystemExit) as excinfo:
-                convert(
-                    "tests/data/convert/invalid.yaml",
-                    "$BASE_URL",
-                    "scanapi.yaml",
-                )
+                settings["input_path"] = "tests/data/convert/invalid.yaml"
+                settings["base_url"] = "$BASE_URL"
+                settings["output_path"] = "scanapi.yaml"
+                openapi_to_scanapi()
 
             assert excinfo.type == SystemExit
             assert excinfo.value.code == 1
@@ -106,11 +120,10 @@ class TestConvert:
     def test_handles_invalid_json_file(self, caplog):
         with caplog.at_level(logging.ERROR):
             with raises(SystemExit) as excinfo:
-                convert(
-                    "tests/data/convert/invalid.json",
-                    "$BASE_URL",
-                    "scanapi.yaml",
-                )
+                settings["input_path"] = "tests/data/convert/invalid.json"
+                settings["base_url"] = "$BASE_URL"
+                settings["output_path"] = "scanapi.yaml"
+                openapi_to_scanapi()
 
             assert excinfo.type == SystemExit
             assert excinfo.value.code == 1
@@ -120,9 +133,10 @@ class TestConvert:
     def test_handles_empty_json_file(self, caplog):
         with caplog.at_level(logging.ERROR):
             with raises(SystemExit) as excinfo:
-                convert(
-                    "tests/data/convert/empty.json", "$BASE_URL", "scanapi.yaml"
-                )
+                settings["input_path"] = "tests/data/convert/empty.json"
+                settings["base_url"] = "$BASE_URL"
+                settings["output_path"] = "scanapi.yaml"
+                openapi_to_scanapi()
 
             assert excinfo.type == SystemExit
             assert excinfo.value.code == 1
@@ -134,11 +148,10 @@ class TestConvert:
     def test_handles_invalid_openapi_schema(self, caplog):
         with caplog.at_level(logging.ERROR):
             with raises(SystemExit) as excinfo:
-                convert(
-                    "tests/data/convert/no_version_definition.yaml",
-                    "$BASE_URL",
-                    "scanapi.yaml",
-                )
+                settings["input_path"] = "tests/data/convert/no_version_definition.yaml"
+                settings["base_url"] = "$BASE_URL"
+                settings["output_path"] = "scanapi.yaml"
+                openapi_to_scanapi()
 
             assert excinfo.type == SystemExit
             assert excinfo.value.code == 1